[refactor] #133 RemoteNotification Live, Test 분리
