+++ Projects/CoreKit/Sources/Data/Client/RemoteNotificationClient/RemoteNotificationsClient+LiveKey.swift
@@ -5,14 +5,14 @@
 //  Created by 김민호 on 7/5/24.
 //
 
-//import Dependencies
-//import UIKit
-//
-//@available(iOSApplicationExtension, unavailable)
-//extension RemoteNotificationsClient: DependencyKey {
-//  public static let liveValue = Self(
-//    isRegistered: { await UIApplication.shared.isRegisteredForRemoteNotifications },
-//    register: { await UIApplication.shared.registerForRemoteNotifications() },
-//    unregister: { await UIApplication.shared.unregisterForRemoteNotifications() }
-//  )
-//}
+import UIKit
+
+import Dependencies
+
+extension RemoteNotificationsClient: DependencyKey {
+    public static let liveValue = Self(
+        isRegistered: { await UIApplication.shared.isRegisteredForRemoteNotifications },
+        register: { await UIApplication.shared.registerForRemoteNotifications() },
+        unregister: { await UIApplication.shared.unregisterForRemoteNotifications() }
+    )
+}
