+++ Projects/CoreKit/Sources/Data/Client/RemoteNotificationClient/RemoteNotificationsClient+TestKey.swift
@@ -8,31 +8,15 @@
 import UIKit
 import Dependencies
 
-extension DependencyValues {
-  public var remoteNotifications: RemoteNotificationsClient {
-    get { self[RemoteNotificationsClient.self] }
-    set { self[RemoteNotificationsClient.self] = newValue }
-  }
-}
-
-@available(iOSApplicationExtension, unavailable)
-extension RemoteNotificationsClient: DependencyKey {
-  public static let liveValue = Self(
-    isRegistered: { await UIApplication.shared.isRegisteredForRemoteNotifications },
-    register: { await UIApplication.shared.registerForRemoteNotifications() },
-    unregister: { await UIApplication.shared.unregisterForRemoteNotifications() }
-  )
-}
-
 extension RemoteNotificationsClient: TestDependencyKey {
-  public static let previewValue = Self.noop
-  public static let testValue = Self()
+    public static let previewValue = Self.noop
+    public static let testValue = Self()
 }
 
 extension RemoteNotificationsClient {
-  public static let noop = Self(
-    isRegistered: { true },
-    register: {},
-    unregister: {}
-  )
+    public static let noop = Self(
+        isRegistered: { true },
+        register: {},
+        unregister: {}
+    )
 }
