+++ Projects/CoreKit/Sources/Data/Client/RemoteNotificationClient/RemoteNotificationsClient.swift
@@ -9,7 +9,7 @@ import DependenciesMacros
 
 @DependencyClient
 public struct RemoteNotificationsClient {
-  public var isRegistered: @Sendable () async -> Bool = { false }
-  public var register: @Sendable () async -> Void
-  public var unregister: @Sendable () async -> Void
+    public var isRegistered: @Sendable () async -> Bool = { false }
+    public var register: @Sendable () async -> Void
+    public var unregister: @Sendable () async -> Void
 }
